refactor(gateway): extract dispatch and thread_updates from server.rs
Break up the oversized server.rs (3271 → 2138 lines) by moving two
cohesive clusters into their own modules:

- dispatch.rs: the chat-turn engine (execute_user_prompt,
  execute_dom_query, await_model_with_cancel, dispatch_text_message).
- thread_updates.rs: client-bound thread/session update frames
  (send_threads_update, thread_history_messages,
  send_thread_messages_update).

Both are pub(crate) and called from the connection handler in server.rs.
Also drops now-dead imports left behind in server.rs and fixes the
module doc comment that still referenced the removed run_stdio_gateway.
